+++ controller/student/student.go
@@ -57,3 +57,16 @@ func PostProfile(c *gin.Context) {
 
 	c.JSON(http.StatusOK, gin.H{"status": 0})
 }
+
+// GetCounselor 获取辅导员信息
+func GetCounselor(c *gin.Context) {
+	grade := c.DefaultQuery("grade", "17级")
+	college := c.DefaultQuery("college", "新闻学院")
+	var counselors []database.Counselor
+
+	if dbe := database.DB.Where("grade = ? AND college = ?", grade, college).Find(&counselors); dbe.Error != nil {
+		c.AbortWithStatusJSON(http.StatusOK, gin.H{"status": 1, "msg": dbe.Error.Error()})
+		return
+	}
+	c.AbortWithStatusJSON(http.StatusOK, gin.H{"status": 0, "counselors": counselors})
+}
