+++ controller/utilities/index.go
@@ -11,6 +11,5 @@ import (
 func GetIndex(c *gin.Context) {
 	var notifies []database.Notify
 	database.DB.Find(&notifies)
-	// c.HTML(http.StatusOK, "common/index.tmpl", gin.H{"notifies": notifies})
 	c.HTML(http.StatusOK, "common/index.html", gin.H{"notifies": notifies})
 }
