+++ controller/utilities/notify.go
@@ -10,9 +10,9 @@ import (
 // GetNotify 获取通知信息
 func GetNotify(c *gin.Context) {
 	var notify database.Notify
-	if dbe := database.DB.First(&notify, c.PostForm("NotifyID")); dbe.Error != nil {
+	if dbe := database.DB.First(&notify, c.PostForm("notifyID")); dbe.Error != nil {
 		c.AbortWithStatusJSON(http.StatusOK, gin.H{"status": 1, "msg": dbe.Error.Error()})
 		return
 	}
-	c.JSON(http.StatusOK, gin.H{"status": 1, "notify": notify})
+	c.JSON(http.StatusOK, gin.H{"status": 0, "notify": notify})
 }
