+++ resources/views/common/index.html
@@ -9,7 +9,7 @@
 	   <script type="text/javascript" src="resources/webframe/jquery/jquery-1.10.2.min.js"></script>
 	   <script type="text/javascript" src="resources/webframe/bootstrap/js/bootstrap.min.js"></script>
        <script type="text/javascript" src="resources/webframe/jquery/jquery.form/jquery.form.js"></script>
-	   <script src="resources/webframe/layer//layer.js"></script>
+	   <script src="resources/webframe/layer/layer.js"></script>
     </head>
     <style type="text/css">
 	#ullist ul{
